Added Developers Page

+++ client/src/redux/actions/profile.ts
@@ -22,6 +22,35 @@ export const getProfile = () => async (dispatch: Dispatch<AppActions>) => {
   }
 };
 
+// Get all profiles
+export const getAllProfiles = () => async (dispatch: Dispatch<AppActions>) => {
+  dispatch({ type: "SET_LOADING", loading: true });
+  try {
+    const res = await axios.get("/api/profile");
+    dispatch({ type: "GET_ALL_PROFILES", profiles: res.data });
+  } catch (e) {
+    dispatch({
+      type: "GET_PROFILE_FAIL",
+      error: { msg: e.response.data.msg, status: e.response.status },
+    });
+  }
+};
+
+// Get profile by id
+export const getProfileById = (id: string) => async (
+  dispatch: Dispatch<AppActions>
+) => {
+  try {
+    const res = await axios.get(`/api/profile/user/${id}`);
+    dispatch({ type: "GET_PROFILE", profile: res.data });
+  } catch (e) {
+    dispatch({
+      type: "GET_PROFILE_FAIL",
+      error: { msg: e.response.data.msg, status: e.response.status },
+    });
+  }
+};
+
 // Create profile
 export const createProfile = (
   formData: CreateProfileFormData,
@@ -118,14 +147,17 @@ export const addEducation = (
 export const deleteExperience = (id: string) => async (
   dispatch: Dispatch<AppActions>
 ) => {
-  try {
-    const res = await axios.delete(`api/profile/experience/${id}`);
-    dispatch({ type: "UPDATE_PROFILE", profile: res.data });
-  } catch (e) {
-    dispatch({
-      type: "GET_PROFILE_FAIL",
-      error: { msg: e.response.data.msg, status: e.response.status },
-    });
+  if (window.confirm("Are you sure? Delete this experience?")) {
+    try {
+      const res = await axios.delete(`api/profile/experience/${id}`);
+      dispatch({ type: "UPDATE_PROFILE", profile: res.data });
+      dispatch<any>(setAlert("Experience removed", "success"));
+    } catch (e) {
+      dispatch({
+        type: "GET_PROFILE_FAIL",
+        error: { msg: e.response.data.msg, status: e.response.status },
+      });
+    }
   }
 };
 
@@ -133,13 +165,35 @@ export const deleteExperience = (id: string) => async (
 export const deleteEducation = (id: string) => async (
   dispatch: Dispatch<AppActions>
 ) => {
-  try {
-    const res = await axios.delete(`api/profile/education/${id}`);
-    dispatch({ type: "UPDATE_PROFILE", profile: res.data });
-  } catch (e) {
-    dispatch({
-      type: "GET_PROFILE_FAIL",
-      error: { msg: e.response.data.msg, status: e.response.status },
-    });
+  if (window.confirm("Are you sure? Delete this education?")) {
+    try {
+      const res = await axios.delete(`api/profile/education/${id}`);
+      dispatch({ type: "UPDATE_PROFILE", profile: res.data });
+      dispatch<any>(setAlert("Education removed", "success"));
+    } catch (e) {
+      dispatch({
+        type: "GET_PROFILE_FAIL",
+        error: { msg: e.response.data.msg, status: e.response.status },
+      });
+    }
+  }
+};
+
+// Delete Account
+export const deleteAccount = () => async (dispatch: Dispatch<AppActions>) => {
+  if (window.confirm("Are you sure? This can't be undone!")) {
+    try {
+      await axios.delete("api/profile");
+      dispatch({ type: "CLEAR_PROFILE" });
+      dispatch({ type: "ACCOUNT_DELETED" });
+      dispatch<any>(
+        setAlert("Your account has been permanently deleted!", "danger")
+      );
+    } catch (e) {
+      dispatch({
+        type: "GET_PROFILE_FAIL",
+        error: { msg: e.response.data.msg, status: e.response.status },
+      });
+    }
   }
 };
